Simplify ReactFiberHydrationWarning in react-reconciler
Stopped printing non-hydratable nodes in the diff, simplified
the ReactFiberHydrationWarning and ReactDOMHostConfig code.
Removed ReactFiberHostConfig functions:
```
getAllChildHostInstances,
getHostInstanceDisplayStringForHydrationWarning,
getHostInstanceClosingDisplayStringForHydrationWarning,
isHydratableInstance,
```
Added ReactFiberHostConfig functions:
```
getHostInstanceDisplayName,
getHostInstanceProps,
```

Removed type HydratableInstance, it everywhere was aliased
to Instance | TextInstance, and was used as such alias,
but in the custom renderer it used an opaque type,
and this failed the typechecks because the aliasing did not happen.
Alternative solution would be to alias HydratableInstance in the
custom renderer's ReactFiberHostConfig.custom to align with other
renderers. I don't see any potential for HydratableInstance to be
something else than Instance | TextInstance, so I chose to inline
the alias.

Removed type HostInstance, not needed anymore.